+++ frappe-custom/mmm_custom/mmm_custom/patches/v1_0/create_ai_fields.py
@@ -1,0 +1,5 @@
+from mmm_custom.setup import create_ai_fields
+
+
+def execute():
+	create_ai_fields()